refactor: Change the project's directory structure. Move codes into ./lib and tests into ./test. Also add index.js
