+++ test/main.test.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const tmp = require('tmp');
-const main = require('./main.js');
+const main = require('../lib/main.js');
 const jsdom = require('jsdom');
 const {JSDOM} = jsdom;
 const Stream = require('stream');
